+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/langchain
 
+## 2.0.0-beta.120
+
+### Patch Changes
+
+- Updated dependencies [457318b]
+  - ai@6.0.0-beta.120
+
+## 2.0.0-beta.119
+
+### Patch Changes
+
+- Updated dependencies [b59d924]
+  - ai@6.0.0-beta.119
+
 ## 2.0.0-beta.118
 
 ### Patch Changes